azure_functions/blobtrigger/python/pyblobtriggermi/requirements.txt
azure-core==1.26.4
azure-functions==1.14.0
azure-storage-blob==12.16.0
azure-storage-queue==12.6.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==40.0.2
idna==3.4
isodate==0.6.1
portalocker==2.7.0
pycparser==2.21
PyJWT==2.7.0
pywin32==306 ; sys_platform == 'win32'
requests==2.30.0
six==1.16.0
typing_extensions==4.5.0
urllib3==2.0.2
